refactor: Car 패키지 변경
