+++ src/main/java/racingcar/domain/RaceRecord.java
@@ -1,7 +1,5 @@
 package racingcar.domain;
 
-import racingcar.race.Car;
-
 import java.util.Objects;
 
 public class RaceRecord {
