+++ src/test/java/racingcar/race/RaceRecordsTest.java
@@ -2,6 +2,7 @@ package racingcar.race;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
 import racingcar.domain.CarName;
 import racingcar.domain.Position;
 import racingcar.rule.MinimumRule;
