+++ src/test/java/study/ListTest.java
@@ -1,10 +1,6 @@
 package study;
 
 import org.junit.jupiter.api.Test;
-import racingcar.race.Car;
-
-import java.util.ArrayList;
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
